get-linux-source: introduce FETCH_DEPTH env variable
FETCH_DEPTH now controls how to get linux source. If it's 0, the
snapshot is searched for and downloaded if found.  Otherwise git clone
--depth $FETCH_DEPTH is executed, and .git is preserved.

+++ get-linux-source/checkout_latest_kernel.sh
@@ -6,6 +6,9 @@ source $(cd $(dirname $0) && pwd)/../helpers.sh
 
 CWD=$(pwd)
 
+# 0 means just download a snapshot
+FETCH_DEPTH=${FETCH_DEPTH:-0}
+
 echo KERNEL_ORIGIN = ${KERNEL_ORIGIN}
 echo KERNEL_BRANCH = ${KERNEL_BRANCH}
 echo REPO_PATH = ${REPO_PATH}
@@ -29,26 +32,27 @@ if [ ! -d "${REPO_PATH}" ]; then
 	mkdir -p $(dirname "${REPO_PATH}")
 	cd $(dirname "${REPO_PATH}")
 	# attempt to fetch desired bpf-next repo snapshot
-	if [ -n "${SNAPSHOT_URL}" ] && \
+	if [ -n "${SNAPSHOT_URL}" ] && [ "${FETCH_DEPTH}" -eq 0 ] && \
 	   wget -U 'BPFCIBot/1.0 (bpf@vger.kernel.org)' -nv ${SNAPSHOT_URL} && \
 	   tar xf bpf-next-${LINUX_SHA}.tar.gz --totals ; then
 		mv bpf-next-${LINUX_SHA} $(basename ${REPO_PATH})
 	else
-		# but fallback to git fetch approach if that fails
-		mkdir -p $(basename ${REPO_PATH})
-		cd $(basename ${REPO_PATH})
-		git init
-		git remote add bpf-next ${KERNEL_ORIGIN}
-		# try shallow clone first
-		git fetch --depth 32 bpf-next
-		# check if desired SHA exists
-		if ! git cat-file -e ${LINUX_SHA}^{commit} ; then
-			# if not, fetch all of bpf-next; slow and painful
-			git fetch bpf-next
-		fi
-		git reset --hard ${LINUX_SHA}
+	    # but fallback to git fetch approach if that fails
+            git clone --depth ${FETCH_DEPTH} ${KERNEL_ORIGIN} ${REPO_PATH}
+            cd "${REPO_PATH}"
+	    # check if desired SHA exists
+	    if ! git cat-file -e ${LINUX_SHA}^{commit} ; then
+                # if not, fetch all of bpf-next; slow and painful
+                git fetch origin
+            fi
+            git reset --hard ${LINUX_SHA}
+            cd -
 	fi
-	rm -rf ${REPO_PATH}/.git || true
+        if [ "${FETCH_DEPTH}" -eq 0 ]; then
+	    rm -rf ${REPO_PATH}/.git || true
+        fi
 
 	foldable end pull_kernel_srcs
+else
+    echo "${REPO_PATH} directory already exists, will not download kernel sources"
 fi
